requirements.txt
playwright==1.48.0
pyyaml==6.0.2
pydantic==2.10.4
requests==2.32.3
python-dateutil==2.9.0
